Module discord_api::models

Module containing various structures received from the Discord API.

Structures here can be created manually though it is advised to fetch actual data using the methods exposed by the http module (which are rexported to discord_api)

Structs

Invite
A struct representing a Discord guild invite, fetched through discord_api::get_invite

InviteGuild
A struct representing a guild that an invite points to

WelcomeChannel
A channel listed in the welcome screen

WelcomeScreen
The welcome screen presented to a user when they join the guild